Time: 745 ms (5.09%) | Memory: 50.4 MB (26.17%) - LeetSync

+++ 943-sum-of-subarray-minimums/sum-of-subarray-minimums.java
@@ -22,7 +22,7 @@ class Solution {
                     break;
             }
 
-            long curr=(long)(i-left+1)*(long)(right-i+1)*(long)arr[i];
+            long curr=(long)(i-left+1)*(right-i+1)*arr[i];
             res=(res+curr)%mod;
         }
